requirements.txt
# poliastro2
astropy>=5.0,<6
astroquery>=0.3.9
jplephem
matplotlib>=2.0,!=3.0.1
numba>=0.46; python_version < '3.9'
numba>=0.53.0; python_version >= '3.9'
numpy
pandas
plotly>=4.0,<6
pyerfa
scipy>=1.4.0